24_01/01_22:00

+++ app/controllers/favorites_controller.rb
@@ -3,13 +3,18 @@ class FavoritesController < ApplicationController
     book = Book.find(params[:book_id])
     favorite = current_user.favorites.new(book_id: book.id)
     favorite.save
-    redirect_to request.referer
   end
 
   def destroy
-    book = Book.find(params[:book_id])
-    favorite = current_user.favorites.find_by(book_id: book.id)
-    favorite.destroy
-    redirect_to request.referer
+    @book = Book.find(params[:book_id])
+    favorite = current_user.favorites.find_by(book_id: @book.id)
+
+    # favorite が存在するか確認してから destroy メソッドを呼ぶ
+    if favorite
+      favorite.destroy
+    else
+      # おそらくエラーメッセージを設定するか、何か他の処理を行う
+      flash[:alert] = "Favorite not found."
+    end
   end
 end
